Move Philips Hue related stuff to a facade
Add an interface for Hue operations
Move "low-level code" to Hue facade
Remove Philips Hue imports from HueService
Add javadoc to HueService
Change /color endpoints to new pretty format
Update integration test
