+++ src/main/java/ch/wisv/chue/states/ColorState.java
@@ -5,6 +5,9 @@ import com.philips.lighting.model.PHBridge;
 import com.philips.lighting.model.PHLightState;
 import javafx.scene.paint.Color;
 
+import java.util.HashMap;
+import java.util.Map;
+
 /**
  * Color state
  */
@@ -12,8 +15,14 @@ public class ColorState implements HueState {
 
     private Color color;
 
+    private Map<String, Color> lightColors = new HashMap<>();
+
     public ColorState(Color color) {
         this.color = color;
+    }
+
+    public Map<String, Color> getLightColors() {
+        return lightColors;
     }
 
     @Override
@@ -24,6 +33,9 @@ public class ColorState implements HueState {
                     +(int) (color.getRed() * 255), (int) (color.getGreen() * 255), (int) (color.getBlue() * 255), "LCT001");
             lightState.setX(xy[0]);
             lightState.setY(xy[1]);
+
+            lightColors.put(id, color);
+
             bridge.updateLightState(id, lightState, null); // If no bridge response is required then use this simpler form.
         }
     }
