+++ src/test/java/ch/wisv/chue/WebControllerTest.java
@@ -1,21 +1,33 @@
 package ch.wisv.chue;
 
+import ch.wisv.chue.hue.HueFacade;
+import com.philips.lighting.model.PHBridge;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.mockito.Spy;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
+import java.util.Arrays;
+
 import static org.hamcrest.CoreMatchers.is;
+import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 public class WebControllerTest {
     @Mock
-    private HueService hueService;
+    private PHBridge phBridge;
+
+    @Mock
+    private HueFacade hueFacade;
+
+    @Spy
+    private HueService hueService = new HueService();
 
     @InjectMocks
     private WebController webController;
@@ -25,6 +37,10 @@ public class WebControllerTest {
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
+
+        when(hueFacade.getAllLightIdentifiers()).thenReturn(Arrays.asList("1", "2", "3"));
+        when(hueFacade.getBridge()).thenReturn(phBridge);
+        hueService.setHueFacade(hueFacade);
 
         mockMvc = MockMvcBuilders.standaloneSetup(webController).build();
     }
@@ -62,27 +78,31 @@ public class WebControllerTest {
     public void testColorHexAll() throws Exception {
         mockMvc.perform(get("/color/all/ff0000"))
                 .andExpect(status().isOk())
-                .andExpect(content().string(is("Changed colour of lamps (all) to #ff0000")));
+                .andExpect(content().string(
+                        is("Time for some new colours: lamp 1 is now #ff0000, lamp 2 is now #ff0000, lamp 3 is now #ff0000")));
     }
 
     @Test
     public void testColorHexID() throws Exception {
         mockMvc.perform(get("/color/1/0000ff"))
                 .andExpect(status().isOk())
-                .andExpect(content().string(is("Changed colour of lamps (1) to #0000ff")));
+                .andExpect(content().string(
+                        is("Time for some new colours: lamp 1 is now #0000ff")));
     }
 
     @Test
     public void testColorFriendlyAll() throws Exception {
         mockMvc.perform(get("/color/all/red"))
                 .andExpect(status().isOk())
-                .andExpect(content().string(is("Changed colour of lamps (all) to #ff0000")));
+                .andExpect(content().string(
+                        is("Time for some new colours: lamp 1 is now #ff0000, lamp 2 is now #ff0000, lamp 3 is now #ff0000")));
     }
 
     @Test
     public void testColorFriendlyID() throws Exception {
         mockMvc.perform(get("/color/2/blue"))
                 .andExpect(status().isOk())
-                .andExpect(content().string(is("Changed colour of lamps (2) to #0000ff")));
+                .andExpect(content().string(
+                        is("Time for some new colours: lamp 2 is now #0000ff")));
     }
 }
